+++ test/blackbox-tests/test-cases/coq/compose-boot-nostdlib.t/B/b.v
@@ -1,0 +1,4 @@
+Print LoadPath.
+Print Prelude.
+Require Import mytheory.
+Check Hello.